Converting into module
