+++ spritesheet1.txt
@@ -1,4 +1,4 @@
-#NOTE$ OBJECTS FACE LEFT BY DEFAULT
+#NOTE: OBJECTS FACE LEFT BY DEFAULT
 $train-car$
 ////|\\\\
 | [] [] |
@@ -102,5 +102,5 @@ $characterx$
 \ /
  X
 / \
-#$DATA$
+#DATA
 #/\,(),ɔc,[],{},`',<>,pq,bd,NИ,Bᗺ,CƆ,cɔ,Kꓘ,Dᗡ,SƧ,sƨ,Pᑫ,yγ,RЯ,rɿ,EƎ,eɘ,ÓÒ,óò,úù,ÚÙ,«»,◭◮
